requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
alembic